--- a/src/pytest_cov/plugin.py
+++ b/src/pytest_cov/plugin.py
@@ -11,6 +11,11 @@ import pytest
 from coverage.results import display_covered
 from coverage.results import should_fail_under
 
+try:
+    from pytest import TerminalReporter  # noqa: PT013
+except ImportError:
+    from _pytest.terminal import TerminalReporter
+
 from . import CovDisabledWarning
 from . import CovFailUnderWarning
 from . import CovReportWarning
@@ -333,7 +338,8 @@ class CovPlugin:
             # it for unit tests that don't need it
             from coverage.misc import CoverageException
 
-            self.cov_controller.sep(self.cov_report, '=', 'coverage report')
+            tr = TerminalReporter(session.config, self.cov_report)
+            tr.write_sep('=', 'coverage report')
             try:
                 self.cov_total = self.cov_controller.summary(self.cov_report)
             except CoverageException as exc: