This PR improves the formatting of separator lines in coverage reports and enhances the robustness of terminal width handling.

#### Why?
- To conform to the visuals from pytest headings.
- To allow clear separation of this plugin output from the previous section

#### Changes:  
- Add a new section heading for the coverage tests output.
- Make all headings the same width as pytest's headings.
- Change separator character of regular headings from `-` to `_`.