@tsvikas 2025-02-18
Thanks for the feedback! I preferred to get things going so the code and output would be visible when discussing it.

I’ve added an example of the updated output to the PR description for clarity. Let me know if you have any concerns!


Output example (with a terminal width of 100)

Before
```
======================================= test session starts ========================================
platform linux -- Python 3.13.1, pytest-8.3.4, pluggy-1.5.0
benchmark: 5.0.1 (defaults: timer=time.perf_counter disable_gc=False min_rounds=5 min_time=0.000005 max_time=1.0 calibration_precision=10 warmup=False warmup_iterations=100000)
rootdir: /tmp/py-temp
configfile: pyproject.toml
plugins: cov-6.0.0, benchmark-5.0.1, loguru-0.4.0, reverse-1.8.0, clarity-1.0.1
collected 4 items

tests/test_cli.py .F.                                                                        [ 75%]
tests/test_.py .                                                                             [100%]

============================================= FAILURES =============================================
_________________________________________ test_app_version _________________________________________

    def test_app_version() -> None:
        result = runner.invoke(app, ["--version"])
        assert result.exit_code == 0
        assert __version__ in result.stdout
>       assert False
E       assert False

result     = <Result okay>

tests/test_cli.py:22: AssertionError

---------- coverage: platform linux, python 3.13.1-final-0 -----------
Name                    Stmts   Miss  Cover
-------------------------------------------
src/ptemp/__init__.py       3      0   100%
src/ptemp/cli.py           19      0   100%
-------------------------------------------
TOTAL                      22      0   100%

===================================== short test summary info ======================================
FAILED tests/test_cli.py::test_app_version - assert False
=================================== 1 failed, 3 passed in 0.33s ====================================

```

After:
```
======================================= test session starts ========================================
platform linux -- Python 3.13.1, pytest-8.3.4, pluggy-1.5.0
benchmark: 5.0.1 (defaults: timer=time.perf_counter disable_gc=False min_rounds=5 min_time=0.000005 max_time=1.0 calibration_precision=10 warmup=False warmup_iterations=100000)
rootdir: /tmp/py-temp
configfile: pyproject.toml
plugins: cov-6.0.0, benchmark-5.0.1, loguru-0.4.0, reverse-1.8.0, clarity-1.0.1
collected 4 items

tests/test_cli.py .F.                                                                        [ 75%]
tests/test_.py .                                                                             [100%]

============================================= FAILURES =============================================
_________________________________________ test_app_version _________________________________________

    def test_app_version() -> None:
        result = runner.invoke(app, ["--version"])
        assert result.exit_code == 0
        assert __version__ in result.stdout
>       assert False
E       assert False

result     = <Result okay>

tests/test_cli.py:22: AssertionError

========================================= coverage report ==========================================
_________________________ coverage: platform linux, python 3.13.1-final-0 __________________________
Name                    Stmts   Miss  Cover
-------------------------------------------
src/ptemp/__init__.py       3      0   100%
src/ptemp/cli.py           19      0   100%
-------------------------------------------
TOTAL                      22      0   100%

===================================== short test summary info ======================================
FAILED tests/test_cli.py::test_app_version - assert False
=================================== 1 failed, 3 passed in 0.18s ====================================

```