@The-Compiler 2025-02-19
I agree your proposed output looks nicer (and indeed it helps to have a "preview"!), just one nitpick about it: With

```
========================================= coverage report ==========================================
_________________________ coverage: platform linux, python 3.13.1-final-0 __________________________
Name                    Stmts   Miss  Cover
-------------------------------------------
src/ptemp/__init__.py       3      0   100%
src/ptemp/cli.py           19      0   100%
-------------------------------------------
TOTAL                      22      0   100%
```

do the two separators really make sense, given that there will only ever be one coverage report (I think?). I'd have expected something like:

```
========================================= coverage report ==========================================
Name                    Stmts   Miss  Cover
-------------------------------------------
src/ptemp/__init__.py       3      0   100%
src/ptemp/cli.py           19      0   100%
-------------------------------------------
TOTAL                      22      0   100%

platform linux, python 3.13.1-final-0
```

or maybe just

```
========================= coverage: platform linux, python 3.13.1-final-0 ==========================
Name                    Stmts   Miss  Cover
-------------------------------------------
src/ptemp/__init__.py       3      0   100%
src/ptemp/cli.py           19      0   100%
-------------------------------------------
TOTAL                      22      0   100%
```