@tsvikas 2025-02-20
As per request, I moved the heading to `pytest_terminal_summary`.  It seems to fit there better :).

It also helped me notice two small behaviors that I changed (I can revert them if needed):  

* Renamed the heading from "coverage report" to "tests coverage" (since in some cases, it's not a full report but a warning).  
* Adjusted newlines to match pytest's style (one newline after `_` headings, no newlines before/after `=` headings).